backend/requirements-enterprise.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.3
firebase-admin==6.4.0
python-dotenv==1.0.1
pandas==2.2.1
openpyxl==3.1.5
reportlab==4.2.2
scikit-learn==1.4.1.post1
numpy==1.26.4
xgboost==2.0.3
networkx==3.2.1
email-validator==2.2.0
